+++ pkg/auth/webserver_test.go
@@ -1,11 +1,6 @@
 package auth
 
 import (
-	"bytes"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/url"
@@ -16,18 +11,17 @@ import (
 	"github.com/dgrijalva/jwt-go"
 	"github.com/weberc2/auth/pkg/testsupport"
 	pz "github.com/weberc2/httpeasy"
-	pztest "github.com/weberc2/httpeasy/testsupport"
 )
 
 func TestLoginHandler(t *testing.T) {
 	now := time.Date(1988, 9, 3, 0, 0, 0, 0, time.UTC)
 	codes := TokenFactory{
-		Issuer:           "issuer",
-		Audience: "audience",
-		TokenValidity:    time.Minute,
-		ParseKey:         []byte("signing-key"),  // symmetric
-		SigningKey:       []byte("signing-key"),  // symmetric
-		SigningMethod:    jwt.SigningMethodHS512, // symmetric, deterministic
+		Issuer:        "issuer",
+		Audience:      "audience",
+		TokenValidity: time.Minute,
+		ParseKey:      []byte("signing-key"),  // symmetric
+		SigningKey:    []byte("signing-key"),  // symmetric
+		SigningMethod: jwt.SigningMethodHS512, // symmetric, deterministic
 	}
 
 	for _, testCase := range []struct {
@@ -102,137 +96,9 @@ func TestLoginHandler(t *testing.T) {
 	}
 }
 
-func TestExchange(t *testing.T) {
-	now := time.Date(1988, 9, 3, 0, 0, 0, 0, time.UTC)
-	jwt.TimeFunc = func() time.Time { return now }
-	defer func() { jwt.TimeFunc = time.Now }()
-	codes := TokenFactory{
-		Issuer:           "issuer",
-		Audience: "audience",
-		TokenValidity:    time.Minute,
-		ParseKey:         []byte("signing-key"),  // symmetric
-		SigningKey:       []byte("signing-key"),  // symmetric
-		SigningMethod:    jwt.SigningMethodHS512, // symmetric, deterministic
-	}
-
-	for _, testCase := range []struct {
-		name         string
-		username     string
-		code         string
-		wantedStatus int
-		wantedBody   pztest.WantedData
-	}{{
-		name:         "good code",
-		username:     "adam",
-		code:         mustString(codes.Create(now, "adam")),
-		wantedStatus: http.StatusOK,
-		wantedBody:   AnyTokens{}, // good enough
-	}} {
-		t.Run(testCase.name, func(t *testing.T) {
-			accessKey, err := ecdsa.GenerateKey(elliptic.P521(), rand.Reader)
-			if err != nil {
-				t.Fatalf("unexpected error generating access key: %v", err)
-			}
-			refreshKey, err := ecdsa.GenerateKey(elliptic.P521(), rand.Reader)
-			if err != nil {
-				t.Fatalf("unexpected error generating refresh key: %v", err)
-			}
-
-			webServer := WebServer{
-				AuthService: AuthService{
-					Notifications: testsupport.NotificationServiceFake{},
-					TokenDetails: TokenDetailsFactory{
-						AccessTokens: TokenFactory{
-							Issuer:           "issuer",
-							Audience: "audience",
-							TokenValidity:    15 * time.Minute,
-							ParseKey:         &accessKey.PublicKey,
-							SigningKey:       accessKey,
-							SigningMethod:    jwt.SigningMethodES512,
-						},
-						RefreshTokens: TokenFactory{
-							Issuer:           "issuer",
-							Audience: "audience",
-							TokenValidity:    7 * 24 * time.Hour,
-							ParseKey:         &refreshKey.PublicKey,
-							SigningKey:       refreshKey,
-							SigningMethod:    jwt.SigningMethodES512,
-						},
-						TimeFunc: func() time.Time { return now },
-					},
-					Codes:    codes,
-					TimeFunc: func() time.Time { return now },
-				},
-				BaseURL:                 "https://auth.example.org",
-				RedirectDomain:          "app.example.org",
-				DefaultRedirectLocation: "https://app.example.org/default/",
-			}
-
-			data, err := json.Marshal(struct {
-				Code string `json:"code"`
-			}{
-				testCase.code,
-			})
-			if err != nil {
-				t.Fatalf("unexpected error marshaling auth code: %v", err)
-			}
-
-			rsp := webServer.Exchange(
-				pz.Request{Body: bytes.NewReader(data)},
-			)
-
-			if rsp.Status != testCase.wantedStatus {
-				data, err := readAll(rsp.Data)
-				if err != nil {
-					t.Logf("reading response body: %v", err)
-				}
-				t.Logf("response body: %s", data)
-
-				data, err = json.Marshal(rsp.Logging)
-				if err != nil {
-					t.Logf("marshaling response logging: %v", err)
-				}
-				t.Logf("response logging: %s", data)
-				t.Fatalf(
-					"Response.Status: wanted `%d`; found `%d`",
-					testCase.wantedStatus,
-					rsp.Status,
-				)
-			}
-
-			if err := pztest.CompareSerializer(
-				testCase.wantedBody,
-				rsp.Data,
-			); err != nil {
-				t.Fatal(err)
-			}
-		})
-	}
-}
-
 func mustString(s string, err error) string {
 	if err != nil {
 		panic(err)
 	}
 	return s
-}
-
-type AnyTokens struct{}
-
-func (AnyTokens) CompareData(data []byte) error {
-	var details TokenDetails
-
-	if err := json.Unmarshal(data, &details); err != nil {
-		return fmt.Errorf("unmarshaling `TokenDetails`: %w", err)
-	}
-
-	if details.AccessToken == "" {
-		return fmt.Errorf("missing access token")
-	}
-
-	if details.RefreshToken == "" {
-		return fmt.Errorf("missing refresh token")
-	}
-
-	return nil
 }
